libcfnet/bwlimit: code style improvements, types, comments
Several points, after the code revision. Still, code will act exactly
as before this commit.

Edward Welbourne wrote:
> Style: open brace on its own line, not at end of previous line.
>
> Indentation and line-splitting is odd.  Please put the && at the end
> of the previous line, so the next line's bwlimit_next lines up under
> that of the previous line, making it easier to see what they have in
> common and where they differ.  Also, the part after || should be lined
> up under the part before; if it looks to you like it does, check for
> tabs (please use spaces for indent); it doesn't appear lined up in the
> code-review !
>
> This 1000000000L is kinda hard to read.  ...
>
> next and lock can (so should) be static; they're only referenced from
> this file.

Dimitris Apostolou wrote:
> maybe you can replace the u_long type with C89/C99 type?

+++ cf-agent/cf-agent.c
@@ -1037,8 +1037,8 @@ static void KeepControlPromises(EvalContext *ctx, const Policy *policy)
         double bval;
         if (DoubleFromString(value, &bval))
         {
-            bwlimit_kbytes = (u_long) ( bval / 1000.0);
-            Log(LOG_LEVEL_VERBOSE, "Setting rate limit to %ld kBytes/sec", bwlimit_kbytes);
+            bwlimit_kbytes = (uint32_t) ( bval / 1000.0);
+            Log(LOG_LEVEL_VERBOSE, "Setting rate limit to %d kBytes/sec", bwlimit_kbytes);
         }
     }
     Nova_Initialize(ctx);
